+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See README.md before updating this file.
 
 <!-- ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.3...master) -->
 * Add denormalized views over traffic tables.
+* Add error logging for NewRelic when @tracker.track or @tracker.save are rescued
 
 ## v1.1.2 [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.2...v1.1.3)